+++ src/Oro/Bundle/EntityConfigBundle/Tests/Functional/DataFixtures/LoadAttributeData.php
@@ -4,8 +4,9 @@ namespace Oro\Bundle\EntityConfigBundle\Tests\Functional\DataFixtures;
 
 use Doctrine\Common\DataFixtures\AbstractFixture;
 use Doctrine\Common\Persistence\ObjectManager;
+use Oro\Bundle\EntityConfigBundle\Config\ConfigManager;
+use Oro\Bundle\EntityConfigBundle\Entity\FieldConfigModel;
 use Oro\Bundle\EntityConfigBundle\Tests\Functional\Environment\AddAttributesToTestActivityTargetMigration;
-use Oro\Bundle\EntityConfigBundle\Tests\Functional\Environment\UpdateAttributesForTestActivityTargetMigration;
 use Oro\Bundle\TestFrameworkBundle\Entity\TestActivityTarget;
 use Symfony\Component\DependencyInjection\ContainerAwareInterface;
 use Symfony\Component\DependencyInjection\ContainerAwareTrait;
@@ -17,18 +18,13 @@ class LoadAttributeData extends AbstractFixture implements ContainerAwareInterfa
 {
     use ContainerAwareTrait;
 
-    const ENTITY_CONFIG_MODEL       = TestActivityTarget::class;
-    const SYSTEM_ATTRIBUTE_1        = AddAttributesToTestActivityTargetMigration::SYSTEM_ATTRIBUTE_1;
-    const SYSTEM_ATTRIBUTE_2        = AddAttributesToTestActivityTargetMigration::SYSTEM_ATTRIBUTE_2;
-    const DELETED_SYSTEM_ATTRIBUTE  = AddAttributesToTestActivityTargetMigration::DELETED_SYSTEM_ATTRIBUTE;
-    const REGULAR_ATTRIBUTE_1       = AddAttributesToTestActivityTargetMigration::REGULAR_ATTRIBUTE_1;
-    const REGULAR_ATTRIBUTE_2       = AddAttributesToTestActivityTargetMigration::REGULAR_ATTRIBUTE_2;
-    const NOT_USED_ATTRIBUTE        = UpdateAttributesForTestActivityTargetMigration::NOT_USED_ATTRIBUTE;
-    const DELETED_REGULAR_ATTRIBUTE = AddAttributesToTestActivityTargetMigration::DELETED_REGULAR_ATTRIBUTE;
+    const ENTITY_CONFIG_MODEL = TestActivityTarget::class;
+    const SYSTEM_ATTRIBUTE_1  = AddAttributesToTestActivityTargetMigration::SYSTEM_ATTRIBUTE_1;
+    const SYSTEM_ATTRIBUTE_2  = AddAttributesToTestActivityTargetMigration::SYSTEM_ATTRIBUTE_2;
+    const REGULAR_ATTRIBUTE_1 = AddAttributesToTestActivityTargetMigration::REGULAR_ATTRIBUTE_1;
+    const REGULAR_ATTRIBUTE_2 = AddAttributesToTestActivityTargetMigration::REGULAR_ATTRIBUTE_2;
 
-    /**
-     * @var array
-     */
+    /** @var array */
     private static $attributesData = [];
 
     /**
@@ -43,17 +39,13 @@ class LoadAttributeData extends AbstractFixture implements ContainerAwareInterfa
         $attributes = [
             self::SYSTEM_ATTRIBUTE_1,
             self::SYSTEM_ATTRIBUTE_2,
-            self::DELETED_SYSTEM_ATTRIBUTE,
             self::REGULAR_ATTRIBUTE_1,
-            self::REGULAR_ATTRIBUTE_2,
-            self::DELETED_REGULAR_ATTRIBUTE,
-            self::NOT_USED_ATTRIBUTE,
+            self::REGULAR_ATTRIBUTE_2
         ];
 
         $configManager = $this->container->get('oro_entity_config.config_manager');
         foreach ($attributes as $attributeName) {
-            $attribute = $configManager->getConfigFieldModel(self::ENTITY_CONFIG_MODEL, $attributeName);
-            self::$attributesData[$attributeName] = $attribute->getId();
+            self::$attributesData[$attributeName] = self::getAttribute($configManager, $attributeName)->getId();
         }
     }
 
@@ -62,8 +54,26 @@ class LoadAttributeData extends AbstractFixture implements ContainerAwareInterfa
      *
      * @return int|null
      */
-    public static function getAttributeIdByName($attributeName)
+    public static function getAttributeIdByName(string $attributeName): ?int
     {
-        return isset(self::$attributesData[$attributeName]) ? self::$attributesData[$attributeName] : null;
+        return self::$attributesData[$attributeName] ?? null;
+    }
+
+    /**
+     * @param ConfigManager $configManager
+     * @param string        $attributeName
+     *
+     * @return FieldConfigModel
+     */
+    public static function getAttribute(ConfigManager $configManager, string $attributeName): FieldConfigModel
+    {
+        $attribute = $configManager->getConfigFieldModel(self::ENTITY_CONFIG_MODEL, $attributeName);
+        if (null === $attribute) {
+            throw new \RuntimeException(
+                sprintf('The attribute "%s::%s" not found.', self::ENTITY_CONFIG_MODEL, $attributeName)
+            );
+        }
+
+        return $attribute;
     }
 }